requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
distro==1.9.0
fastapi==0.135.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
invoke==2.2.1
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.22
langchain-openai==1.1.12
langsmith==0.7.22
openai==2.30.0
orjson==3.11.7
paramiko==4.0.0
passlib==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
PyMySQL==1.1.2
PyNaCl==1.6.2
PySocks==1.7.1
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlmodel==0.0.37
sshtunnel==0.4.0
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
xxhash==3.6.0
zstandard==0.25.0
